requirements.txt
# 🔗 Core LangChain packages
langchain==0.2.1
langchain-core==0.13.0
langchain-community==0.2.0
langchain-google-genai==0.0.10

# 🔍 Vector store and Embedding
faiss-cpu==1.8.0
google-generativeai==0.5.0

# 📚 LLM Support & OpenAI-compatible types
openai==1.30.1

# 🧠 LLM client for LangChain
google-api-core==2.17.1
google-auth==2.29.0
google-cloud-aiplatform==1.48.0

# 🌐 Web search tool
duckduckgo-search==4.4

# 🧪 Utility Libraries
tqdm==4.66.4
python-dotenv==1.0.1

# 📦 API Server
fastapi==0.111.0
uvicorn==0.30.1

# 🛠️ Additional dependencies
pydantic==2.7.1
numpy==1.26.4
